requirements.txt
pandas==1.5.0
pyairtable==1.3.0
python-dotenv==0.21.0
python_dateutil==2.8.2
streamlit==1.14.0
